tests/wpt/tests/css/css-conditional/container-queries/display-contents-dynamic-style-queries.html

<!doctype html>
<title>CSS Container Queries Test: Invalidate style queries and display:contents</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#style-container">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/cq-testcommon.js"></script>
<style>
  #container.contents {
    --foo: bar;
    display: contents;
  }
  #target {
    color: red;
  }
  @container style(--foo: bar) {
    #target {
      color: green;
    }
  }
</style>
<div id="container">
  <div id="target">This should be green</div>
</div>
<script>
  setup(() => assert_implements_style_container_queries());

  test(() => {
    assert_equals(getComputedStyle(target).color, "rgb(255, 0, 0)");
  }, "Initially the color is red");

  test(() => {
    container.className = "contents";
    assert_equals(getComputedStyle(target).color, "rgb(0, 128, 0)");
  }, "After display and --foo changes, style() query causes the color to be green");
</script>
